Fix bug with ID file that has mixed lane/sample IDs
This (hopefully) fixes the problem highlighted in ticket #474902,
which is caused by running "generate_sample_manifest" with an input file
that contains both sample and lane IDs. The issue was that if the first
line of the file was a sample ID rather than a lane ID, the script would
generate a spreadsheet with "not found" for most IDs.

This commit fixes the problem by assuming that the file contains
entirely lane IDs or entirely sample IDs. It then requires the user to
specify the type of IDs in their input file, using the "--file_id_type"
flag, as per the pathfind script, and defaults to "lane".